withapi/requirements.txt
requests == v2.28.2
Pillow == v9.5.0